Move extract command into own file
