+++ cli.js
@@ -3,12 +3,7 @@
 
 Error.stackTraceLimit = 300;
 
-const path = require('path');
-const fs = require('fs');
-
-const mkdirp = require('mkdirp');
 const yargs = require('yargs');
-const wbn = require('wbn');
 /* const ethereumjs = {
   Tx: require('ethereumjs-tx').Transaction,
 }; */
@@ -36,7 +31,7 @@ try {
   console.warn(err);
 } */
 
-let handled = false;
+const handled = false;
 yargs
   .scriptName('xrpk')
   .command(require('./commands/whoami'))
@@ -62,29 +57,7 @@ yargs
   .command(require('./commands/cat'))
   .command(require('./commands/icon'))
   .command(require('./commands/headers'))
-  .command('extract [input]', 'extract contents of .wbn file', yargs => {
-    yargs
-      .positional('input', {
-        describe: '.wbn file to extract',
-        // default: 5000
-      });
-  }, async argv => {
-    handled = true;
-
-    if (argv.input) {
-      const d = fs.readFileSync(argv.input);
-      const bundle = new wbn.Bundle(d);
-      for (const url of bundle.urls) {
-        const pathname = new URL(url).pathname.slice(1);
-        console.log(pathname);
-        const dirname = path.dirname(pathname);
-        await mkdirp(dirname);
-        fs.writeFileSync(pathname, bundle.getResponse(url).body);
-      }
-    } else {
-      console.warn('missing input file');
-    }
-  })
+  .command(require('./commands/extract'))
   .showHelpOnFail(false)
   .argv;
 if (!handled) {
